+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
-## 0.0.1
+## 0.1.0
 
-- Initial version, created by Stagehand
+  - Initial version, includes a `futureMiddleware` that intercepts Futures and FutureActions, captures their results, and dispatches their results as Redux actions.
+  - Included classes
+    - `FutureAction`
+    - `FutureFulfilledAction`
+    - `FutureRejectedAction`